cc: redesign IR pass interface, migrate shifts + Float16 to mapping pass
Add Function::create_reg_pseudo() and Instruction::call_with_abi() as
canonical IR primitives for passes that synthesize instructions. These
replace duplicated patterns across mapping.rs (alloc_reg64), lower.rs
(open-coded pseudo alloc), and the 6+ copies of the ABI-classify-and-
build-call sequence (build_rtlib_call, build_rtlib_call_explicit,
RtlibCallParams — all deleted).

Migrate int128 constant shifts (Shl/Lsr/Asr) from backend codegen to
the mapping pass as expand_int128_const_shl/lsr/asr. Variable shifts
remain in backends (require arch-specific branching).

Migrate Float16 conversions from linearizer (emit_float16_convert_call)
to the x86_64 mapping pass. The linearizer now emits standard FCvtF/
FCvtS/FCvtU/SCvtF/UCvtF ops; the mapper lowers Float16 variants to
rtlib calls with proper compiler-rt vs libgcc ABI dispatch.

Fix uint128 large constant sign-extension: (*v >> 64) as i64 sign-
extends for values like 0xFFFFFFFFFFFFFFFF; change to as u64 as i64
in three x86_64 locations.

+++ cc/arch/codegen.rs
@@ -284,7 +284,7 @@ impl<I: LirInst + EmitAsm> CodeGenBase<I> {
                 if size > 8 {
                     // 128-bit: emit as two quads (little-endian: lo then hi)
                     let lo = *val as i64;
-                    let hi = (*val >> 64) as i64;
+                    let hi = (*val >> 64) as u64 as i64;
                     self.push_directive(Directive::Quad(lo));
                     self.push_directive(Directive::Quad(hi));
                 } else {
