+++ cc/arch/x86_64/codegen.rs
@@ -2846,7 +2846,7 @@ impl X86_64CodeGen {
         match &src_loc {
             Loc::Imm(v) => {
                 let lo = *v as i64;
-                let hi = (*v >> 64) as i64;
+                let hi = (*v >> 64) as u64 as i64;
                 // Store lo half
                 if lo > i32::MAX as i64 || lo < i32::MIN as i64 {
                     self.push_lir(X86Inst::MovAbs {
